backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
anthropic>=0.43
openai>=1.59
google-genai>=1.10
python-dotenv>=1.0
pydantic>=2.10
